# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: RelayFlow API
  version: 0.1.0
  description: API contract for the RelayFlow backend.
servers:
  - url: http://localhost:8080
paths:
...
  /auth/bootstrap:
    post:
      operationId: bootstrap
      summary: Provision the first admin account and workspace on a fresh instance
      description: >
        Public, unauthenticated endpoint that establishes a session on success. Only
        succeeds when no user accounts exist yet; returns 409 once the instance has
        already been bootstrapped.
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BootstrapRequest"
      responses:
        "200":
          description: Instance bootstrapped; admin account, workspace, and session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapResponse"
        "409":
          description: This instance has already been set up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/signup:
    post:
      operationId: signup
      summary: Register a new account by accepting a workspace invite
      description: >
        Invite-gated signup — there is no open public registration. The invite token
        is validated with the same rules as `POST /invites/{token}/accept` (email
        match, not revoked/expired) and the new user is added to the invite's
        workspace as part of account creation. Establishes a session on success.
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "200":
          description: Account created, invite accepted, and session established
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupResponse"
        "404":
          description: Invite not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: An account with this email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "410":
          description: Invite has been revoked or has expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invite was sent to a different email address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      summary: Log in with email and password
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session established, or a two-factor challenge returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedUserResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login/2fa:
    post:
      operationId: login2FA
      summary: Complete a two-factor login challenge
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Login2FARequest"
      responses:
        "200":
          description: Session established after OTP verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedUserResponse"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
...
  schemas:
...
    BootstrapRequest:
      type: object
      required: [name, email, password, workspaceName]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        workspaceName:
          type: string
    SignupRequest:
      type: object
      required: [name, email, password, inviteToken]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        inviteToken:
          type: string
          format: uuid
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    Login2FARequest:
      type: object
      required: [challengeToken, otp]
      properties:
        challengeToken:
          type: string
        otp:
          type: string
...
    ErrorResponse:
      type: object
      required: [message, timestamp]
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    AuthenticatedUserResponse:
      type: object
      required: [authenticated, twoFactorRequired]
      properties:
        authenticated:
          type: boolean
        userId:
          type: string
          format: uuid
          nullable: true
        email:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        twoFactorRequired:
          type: boolean
          description: True when password login succeeded but OTP verification is still required.
        challengeToken:
          type: string
          nullable: true
          description: Short-lived token passed to `/auth/login/2fa`.
...
    BootstrapResponse:
      type: object
      required: [authenticated, userId, email, displayName, workspaceId]
      properties:
        authenticated:
          type: boolean
        userId:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        workspaceId:
          type: string
          format: uuid
    SignupResponse:
      type: object
      required: [authenticated, userId, email, displayName, workspaceId]
      properties:
        authenticated:
          type: boolean
        userId:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        workspaceId:
          type: string
          format: uuid
...